Potential fix for code scanning alert no. 45: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger type before the result is assigned to the size_t variable. This can be done by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done using the larger type, and any potential overflow will be avoided.

Specifically, we will cast preds_w to size_t before multiplying it by preds_h on line 153.

+++ libwebp-1.0.3/src/enc/webp_enc.c
@@ -150,7 +150,7 @@ static VP8Encoder* InitVP8Encoder(const WebPConfig* const config,
   const int mb_h = (picture->height + 15) >> 4;
   const int preds_w = 4 * mb_w + 1;
   const int preds_h = 4 * mb_h + 1;
-  const size_t preds_size = preds_w * preds_h * sizeof(*enc->preds_);
+  const size_t preds_size = (size_t)preds_w * preds_h * sizeof(*enc->preds_);
   const int top_stride = mb_w * 16;
   const size_t nz_size = (mb_w + 1) * sizeof(*enc->nz_) + WEBP_ALIGN_CST;
   const size_t info_size = mb_w * mb_h * sizeof(*enc->mb_info_);
